Added a testing program for the Endterm assignment 1
Added a testing program for the Endterm assignment 1.

This program can be used to test if you implementation is as it will be expected in the OODP endterm exam. This program has been build to check all the code in the same way as CodeGrade does and will just like CodeGrade return the correct score for the assignment